Switched to 32-bit integers

+++ Parallel/parBubbleSort.cpp
@@ -7,7 +7,7 @@
 #include "parSorts.hpp"
 
 
-void parBubbleSort(std::vector<int64_t>* arr)
+void parBubbleSort(std::vector<int32_t>* arr)
 {
 	
 }
